+++ URL-Ultimate-Filter-Surge.js
@@ -1,14 +1,14 @@
 /**
  * @file    URL-Ultimate-Filter-Surge.js
- * @version 45.44
+ * @version 45.45
  * @date    2026-04-23
- * @rules   1424 total (290 domains, 351 critical paths, 403 path keywords, 109 param rules)
+ * @rules   1431 total (290 domains, 358 critical paths, 403 path keywords, 109 param rules)
  * @build   SSOT Compiler — Dual-Target Compilation
  */
 
 const CONFIG = { DEBUG_MODE: false, AC_SCAN_MAX_LENGTH: 600 };
-const SCRIPT_VERSION = '45.44';
-const SCRIPT_BUILD = 'V45.44 (2026-04-23) | 1424 rules | 2825 tests';
+const SCRIPT_VERSION = '45.45';
+const SCRIPT_BUILD = 'V45.45 (2026-04-23) | 1431 rules | 2846 tests';
 const EMPTY_SET = new Set();
 
 const OAUTH_SAFE_HARBOR = {
@@ -643,6 +643,21 @@ const RULES = {
       ])],
     ['ntm.pstatic.net', new Set([
         'DROP:/'
+      ])],
+    ['fp.amap.com', new Set([
+        'DROP:/ws/shield/location/fp/report'
+      ])],
+    ['awaken.amap.com', new Set([
+        'DROP:/ws/h5_log'
+      ])],
+    ['m5.amap.com', new Set([
+        'DROP:/ws/shield/nest/updatable/v1/log', 'DROP:/ws/feature/preheat/bootevent', '/ws/valueadded/alimama/splash_screen'
+      ])],
+    ['m5-zb.amap.com', new Set([
+        'DROP:/ws/security/account/device_reporting'
+      ])],
+    ['m5-x.amap.com', new Set([
+        'DROP:/ws/shield/amapstream/upload'
       ])]
   ])
   },